# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CAASPP & ELPAC Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [student, teacher, admin]
        - name: classId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user details (e.g., assign class)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /classes:
    get:
      operationId: listClasses
      tags: [users]
      summary: List all classes
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [users]
      summary: Create a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{classId}:
    patch:
      operationId: updateClass
      tags: [users]
      summary: Update class details
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassRequest"
      responses:
        "200":
          description: Class updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    delete:
      operationId: deleteClass
      tags: [users]
      summary: Delete a class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Class deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, teacher, admin]
        grade:
          type: string
          nullable: true
        classIds:
          type: array
          items:
            type: string
          nullable: true
        className:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, username, name, email, role, createdAt]
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, teacher, admin]
        grade:
          type: string
          nullable: true
        classIds:
          type: array
          items:
            type: string
          nullable: true
      required: [username, name, email, password, role]
    UpdateUserRequest:
      type: object
      properties:
        classIds:
          type: array
          items:
            type: string
          nullable: true
        grade:
          type: string
          nullable: true
...
    Class:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        grade:
          type: string
        section:
          type: string
          nullable: true
        teacherId:
          type: string
        teacherName:
          type: string
        studentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, grade, teacherId, teacherName, studentCount, createdAt]
    CreateClassRequest:
      type: object
      properties:
        name:
          type: string
        grade:
          type: string
        section:
          type: string
          nullable: true
        teacherId:
          type: string
      required: [name, grade, teacherId]
    UpdateClassRequest:
      type: object
      properties:
        name:
          type: string
        grade:
          type: string
        section:
          type: string
          nullable: true
        teacherId:
          type: string