# lib/api-zod/src/generated/types/updateClassRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAASPP & ELPAC Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateClassRequest {
  name?: string;
  grade?: string;
  section?: string | null;
  teacherId?: string;
}